Add GameLog class
Because this game is designed to be played on a single computer, with
players alternating turns, and because they should not be watching as
the other player makes their turn (else they might see the other
player's hand), each player needs to be able to review the moves made by
the previous player during their own turn. For this reason I have added
the GameLog class to store all moves made during the game so they can be
recalled on command.