Finish writing game logic
With this commit the game can be considered v1.0. All of the core logic
has been implemented, and all that remains to do is clean up the code and
implement small QoL changes.

This commit also overhauled the GameLog class once again, in the process
adding a substantial amount of ugly code within the larger codebase in
order to use it correctly. This change was necessary in order to assure
that the log was properly recalling each important game move (especially
Just Say No plays, which would not have been possible with the old
version), but there has got to be a better, more elegant way to solve
this problem. It's a work in progress.